integration tests and switch to percentage based configs

+++ lib/src/payloads.rs
@@ -4,7 +4,7 @@ use serde::{Deserialize, Serialize};
 pub struct StatusPayload {
     pub enabled: bool,
     pub gamma_state: String,
-    pub gamma: u8,
+    pub gamma: f32,
 }
 #[derive(Serialize, Deserialize, Debug)]
 pub struct TogglePayload {
@@ -19,5 +19,5 @@ pub struct ErrorPayload {
 
 #[derive(Serialize, Deserialize, Debug)]
 pub struct SetPayload {
-    pub gamma: u8,
+    pub gamma: f32,
 }
